+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `navigator.getInstalledRelatedApps()` shim
+
 ## [v2.5.1] - 2020-08-21
 
 ### Added