+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `navigator.getInstalledRelatedApps()` shim
+- Add `UTM` generator and validator class, extending `URL`
 
 ## [v2.5.1] - 2020-08-21