+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Promise.prototype.finally` shim
 - `preload()` function, creating a `<link rel="preload">`
 - Add `parseHTML()` function, which parses text and returns a document fragment
+- More functions for other preloading techniques
 
 ## [v2.5.1] - 2020-08-21